backend: split app/api/routes.py into domain submodules
Previously a single 394-line module held all REST + WebSocket endpoints.
Split into app/api/routes/ package with one module per concern:

- lifespan.py: background session cleanup, per-OID init locks
- session.py: POST /session/init
- state.py:   GET /state, /config
- game.py:    POST /game/* (point, set, timeout, serve, score, reset)
- display.py: POST /display/* (visibility, simple-mode)
- customization.py: GET/PUT /customization
- overlays.py: GET /overlays, /teams, /themes, /links, /styles
- websocket.py: WebSocket /ws

__init__.py re-exports api_router and router_lifespan so existing
imports ("from app.api.routes import api_router") keep working.

Mirrors the app/overlay/overlay_backends/ split pattern from PR #147.
